Fiverr 4th order
The necessary packages (tidyverse, magrittr, summarytools, rstatix) are installed and loaded.

The current working directory is set to "C:/R DATA SETS".

The HELP dataset is read from a CSV file called "HELP.csv" and stored in the 'hdata' variable. The first few rows of the dataset are displayed using the 'head' function.

The column names are attached for easy access using the 'attach' function.

The frequency distributions of the 'sex' and 'd1' variables are examined using the 'freq' function. The number of female patients is 107, and the number of patients who have never been hospitalized is 92.

The percentage of patients who have been hospitalized fewer than 5 times is calculated as 62.91%.

The dataset is subsetted to include only patients whose primary substance of abuse is cocaine and who are at least 40 years old. The resulting subset is stored in the 'df' variable.

The frequency distribution of age is examined for the subsetted data, and it is found that there are 27 patients who are at least 40 years old and whose primary substance of abuse is cocaine.

A frequency table of sex based on the subset is created, and it is determined that there are 7 female patients and 20 male patients among those who are at least 40 years old and have cocaine listed as their primary substance of abuse.

Descriptive statistics for the variable 'e2b', which represents the number of times the respondent entered a detox program in the past 6 months, are examined. The mean is 2.50, the median is 2.00, and the maximum value is 21.00.

A new variable called 'Mdays' is created by taking the mean of the 'daysanysub' and 'dayslink' variables. Descriptive statistics for the new variable are examined, revealing a mean of 159.26, a minimum value of 7.00, and a maximum value of 358.50.

Overall, the code provides frequency distributions, descriptive statistics, and subsets of the HELP dataset to analyze various aspects of the data, including patient characteristics, substance abuse, and treatment programs.